feat: Remove version and createdAt in favor of lastModifiedAt
I was planning on using etags with an @Version property, however, I decided to use Last-Modified/If-Unmodified-Since headers. The downside of using timestamps is they are only down to the second, which could cause issues in very high-volume applications. This is eventually going to be a multi-tenant application where each user has their own records. Given that fact, I don't think the same user will save the same record twice within 1 second.

Also, it could be useful to display the Last Modified At data in the UI. When using websites I sometimes like to see the Last Modified Timestamp, whereas seeing an @Version number in the UI wouldn't be useful.