Improve exceptions
Change exceptions to extend ResponseStatusException instead of RuntimeException. The benefit is, the message passed into the exception constructor is displayed in the response. This means I can't have different error message responses each time the exception is throw instead of a generic message that applies to all exceptions of that type.