Use Unmodifiable Lists in Domain Layer
Previously I was holding lists as Stream<T> in the domain layer. This was a way to signal to the caller "Hey this is immutable" via the Type system; however, Streams have a major limitation in that they can only be used once.

For now, I've fallen back to using Collections.unmodifiableList() which has the same result of being unmodifiable, the only problem is error occur at run time not compile time and there is no signal from the type system that the list is unmodifiable. I have to resort to using javadocs.

This is not a great solution. I will continue to think about how to improve this.