Make Category static
Having dynamic categories was causing too much headache in the business logic for not enough benefits.

I have replaced the CategoryEntity with a Category enum.

May consider making categories dynamic at some point in the future.

This change has made the lineItems.type field a bit redundant, but I am going to leave it for now.