Simplify Database Schema
I removed the two domain tables in favor of just using enums in code. One thing I did not want to lose was case safety when reading/writing to those columns. Luckily hibernate has @ColumnTransformer for that. So when reading/writing from the DB the @ColumnTransformer tells hibernate to uppercase the values. This means even if the record has 'income' in the database it will correctly serialize in to an ItemType enum.